Video list and Album list merging
Added code to merge video list within the album list structure. This
will lead to the ability to use the localStorage version when the video
page is displayed. Also it can serve as a container for tracking the
Caching of individual videos. The cache JS code is in a separate file:
wvbs-cache.js